Added blind layer, mainly for position encoding
A blind layer needs "no" input. However, the input array's second
dimension is used to communicate desired number of positions. This
may be unconventional and change in the near future. It is a
requirement to building the self-attention-head correctly, though.

One large missunderstanding: KQT was calculated in a way that
builds a matrix from entries of K and Q vectors. It SHOULD have
been KQT (K x Q^T), which builds a matrix of dot-products of K and
Q with dim of time-entries. Still needs fixing.

Some minior fixes, e.g. allocating 'line' in network_training.f08